+++ Changes.md
@@ -1,3 +1,4 @@
+* #1202 - Add gear icons to rubrics when referenced by a virtual category
 * #1170 - Separate out finalist non-numeric nominees
 * #1198 - Add support for virtual subjective categories
 * #1200 - Remove normalization in favor of scaling